+++ src/app/models/casoevidencia.ts
@@ -4,4 +4,5 @@ import { Evidencia } from './evidencia';
 export interface CasoConEvidencia {
     caso: Caso;
     ultimaEvidencia: Evidencia | null;
+    rutsUnicos?: string[];
 }